Feat: max gas issuance type change (#17)
* feat: max_gas_issuance_per_block type changed to u32

* fix: return cuda for risc0

+++ lib/src/input/pacaya.rs
@@ -9,7 +9,7 @@ sol! {
         uint8 sharingPctg;
         uint32 gasIssuancePerSecond;
         uint64 minGasExcess;
-        uint64 maxGasIssuancePerBlock;
+        uint32 maxGasIssuancePerBlock;
     }
 
     #[derive(Debug, Default, Deserialize, Serialize)]
